feat: check logged in/production env in middleware
